update UML and type refactor

+++ GameManager.cpp
@@ -10,21 +10,21 @@
 #include "Humain.hpp"
 #include "Vampire.hpp"
 
-GameManager::GameManager(int width, int height, int humainsNb, int vampiresNb) :
+GameManager::GameManager(size_t width, size_t height, size_t humainsNb, size_t vampiresNb) :
     _width(width), _height(height), _humainsNb(humainsNb), _vampiresNb(vampiresNb)
 {
     // Seed the random with current time to have all the time new random sequence
     srand((unsigned int)time(NULL));
 
-    int randX, randY;
+    size_t randX, randY;
     // Fill all humains with random position
-    for(int i = 0; i < _humainsNb; ++i){
+    for(size_t i = 0; i < _humainsNb; ++i){
         randX = rand() % (_width);
         randY = rand() % (_height);
        _humanoids.push_back(new Humain(randX, randY));
     }
     // Fill all vampires with random position
-    for(int i = 0; i < vampiresNb; ++i){
+    for(size_t i = 0; i < vampiresNb; ++i){
         randX = rand() % (_width);
         randY = rand() % (_height);
         _humanoids.push_back(new Vampire(randX, randY));
@@ -33,10 +33,10 @@ GameManager::GameManager(int width, int height, int humainsNb, int vampiresNb) :
 }
 
 void GameManager::displayGameState() const {
-    std::cout << _cornerBorderChar << std::setfill(_upDownBorderChar) << std::setw(_width + 1) << _cornerBorderChar << std::endl;
-    for(int y = 0; y < _height; ++y){
+    std::cout << _cornerBorderChar << std::setfill(_upDownBorderChar) << std::setw((int)_width + 1) << _cornerBorderChar << std::endl;
+    for(size_t y = 0; y < _height; ++y){
         std::cout << _leftRightBorderChar;
-        for(int x = 0; x < _width; ++x){
+        for(size_t x = 0; x < _width; ++x){
             char charToDisplay = ' ';
             for(Humanoid* humanoid: _humanoids){
                 if(humanoid->getPosX() == x && humanoid->getPosY() == y){
@@ -47,6 +47,6 @@ void GameManager::displayGameState() const {
         }
         std::cout << _leftRightBorderChar << std::endl;
     }
-    std::cout << _cornerBorderChar << std::setfill(_upDownBorderChar) << std::setw(_width + 1) << _cornerBorderChar << std::endl;
+    std::cout << _cornerBorderChar << std::setfill(_upDownBorderChar) << std::setw((int)_width + 1) << _cornerBorderChar << std::endl;
 }
